+++ contracts/cosmwasm-vm/cw-common/Cargo.toml
@@ -14,7 +14,7 @@ cosmwasm-std = { version = "1.2.1", features = ["iterator", "ibc3"] }
 ibc = { version = "0.32.0", default-features = false, features = ["parity-scale-codec", "mocks-no-std", "serde", "borsh"]}
 serde = { version = "1.0.154", default-features = false, features = ["derive"] }
 cw-storage-plus = "1.0.1"
-hex ="0.4.3"
-
+hex = { version = "0.4.3", default-features = false}
+common = { path="../../../libraries/rust/common" }
 [dev-dependencies]
 getrandom = {version = "0.2", default-features = false, features = ["custom"]}
