+++ contracts/cosmwasm-vm/cw-ibc-core/src/gas_estimates.rs
@@ -1,4 +1,5 @@
 pub const GAS_NUMERATOR_DEFAULT: u64 = 4;
 pub const GAS_ADJUSTMENT_NUMERATOR_DEFAULT: u64 = 150;
 pub const GAS_DENOMINATOR: u64 = 100;
-pub const GAS_FOR_SUBMESSAGE_LIGHTCLIENT: u64 = 500_000;
+pub const GAS_FOR_SUBMESSAGE_LIGHTCLIENT: u64 = 1_000_000;
+pub const GAS_FOR_SUBMESSAGE_XCALL: u64 = 900_000;
