+++ contracts/cosmwasm-vm/cw-xcall/src/ack.rs
@@ -1,14 +1,5 @@
 use super::*;
 
-/// IBC ACK. See:
-/// https://github.com/cosmos/cosmos-sdk/blob/f999b1ff05a4db4a338a855713864497bedd4396/proto/ibc/core/channel/v1/channel.proto#L141-L147
-#[derive(Serialize, Deserialize, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub enum Ack {
-    Result(Binary),
-    Error(String),
-}
-
 pub fn make_ack_success() -> Binary {
     let res = Ack::Result(b"1".into());
