+++ contracts/cosmwasm-vm/cw-xcall/src/bin/schema.rs
@@ -1,6 +1,7 @@
 use cosmwasm_schema::write_api;
 
-use cw_xcall::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
+use cw_common::xcall_msg::ExecuteMsg;
+use cw_xcall::msg::{InstantiateMsg, QueryMsg};
 
 fn main() {
     write_api! {
