+++ contracts/cosmwasm-vm/cw-xcall/src/types/mod.rs
@@ -1,18 +1,14 @@
-pub mod address;
 pub mod call_request;
 pub mod message;
 pub mod request;
 pub mod response;
 pub mod storage_keys;
-use address::Address;
-pub use common::rlp;
-use cosmwasm_schema::cw_serde;
-use cw_storage_plus::KeyDeserialize;
-
-use common::rlp::{Decodable, Encodable};
-use cosmwasm_std::{to_binary, Binary};
-use request::CallServiceMessageRequest;
-use response::CallServiceMessageResponse;
-use std::fmt::Display;
 
 use crate::error::ContractError;
+pub use common::rlp;
+use common::rlp::{Decodable, Encodable};
+use cosmwasm_schema::cw_serde;
+use cosmwasm_std::{to_binary, Binary};
+use cw_common::types::Address;
+use request::CallServiceMessageRequest;
+use response::CallServiceMessageResponse;
