+++ contracts/cosmwasm-vm/cw-xcall/tests/test_admin.rs
@@ -2,7 +2,8 @@ mod account;
 mod setup;
 use account::*;
 use cosmwasm_std::{from_binary, testing::mock_env};
-use cw_xcall::{state::CwCallService, types::address::Address};
+use cw_common::types::Address;
+use cw_xcall::state::CwCallService;
 use setup::*;
 
 #[test]
@@ -279,7 +280,7 @@ fn add_invalid_char_as_admin() {
             mock_deps.as_mut(),
             mock_env,
             mock_info.clone(),
-            cw_xcall::msg::ExecuteMsg::SetAdmin {
+            cw_common::xcall_msg::ExecuteMsg::SetAdmin {
                 address: "*************".into(),
             },
         )
@@ -308,7 +309,7 @@ fn update_admin_invalid_chars() {
             mock_deps.as_mut(),
             mock_env.clone(),
             mock_info.clone(),
-            cw_xcall::msg::ExecuteMsg::SetAdmin {
+            cw_common::xcall_msg::ExecuteMsg::SetAdmin {
                 address: admin_one(),
             },
         )
@@ -323,7 +324,7 @@ fn update_admin_invalid_chars() {
             mock_deps.as_mut(),
             mock_env,
             mock_info.clone(),
-            cw_xcall::msg::ExecuteMsg::UpdateAdmin {
+            cw_common::xcall_msg::ExecuteMsg::UpdateAdmin {
                 address: "*****%%%%%@@@###!1234hello".into(),
             },
         )
@@ -354,7 +355,7 @@ fn validate_address_add_admin_size_lessthan_3() {
             mock_deps.as_mut(),
             mock_env,
             mock_info.clone(),
-            cw_xcall::msg::ExecuteMsg::SetAdmin {
+            cw_common::xcall_msg::ExecuteMsg::SetAdmin {
                 address: "sm".into(),
             },
         )
@@ -385,7 +386,7 @@ fn validate_address_add_admin_size_more_than_45() {
             mock_deps.as_mut(),
             mock_env,
             mock_info.clone(),
-            cw_xcall::msg::ExecuteMsg::SetAdmin {
+            cw_common::xcall_msg::ExecuteMsg::SetAdmin {
                 address: "eddiuo6lbp05golmz3rb5n7hbi4c5hhyh0rb1w6cslyjt5mhwd0chn3x254lyorpx4dzvrvsc9h2em44be2rj193dwe".into(),
             },
         )
